% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchBoxplot.R
\name{batchBoxplot}
\alias{batchBoxplot}
\title{Boxplot for Batch Effects}
\usage{
batchBoxplot(
  idvar,
  batchvar,
  feature,
  formula,
  ranef,
  data,
  adjustBatch = FALSE,
  orderby = "mean",
  plotMeans = TRUE,
  colors = "grey",
  xlabel = "batch",
  ylabel = "residuals",
  title = "",
  verbose = TRUE,
  ...
)
}
\arguments{
\item{idvar}{character string that specifies name of ID variable. ID variable can be factor, numeric, or character.}

\item{batchvar}{character string that specifies name of the batch variable. Batch variable should be a factor.}

\item{feature}{character string that specifies name of the numeric feature variable, or the numeric index of the corresponding column.}

\item{formula}{character string representing all fixed effects on the right side of the formula for the linear mixed effects model. This should be in the notation used by \code{lme4} and include covariates, time, and any interactions. For example, \code{"age + sex + diagnosis*time"} fits model with fixed effects age, sex, diagnosis, time, and the diagnosis*time interaction. Formula should NOT include batchvar and should NOT include random effects.}

\item{ranef}{character string representing formula for the random effects in the notation used by \code{lme4}. For example, \code{"(1|subid)"} fits a random intercept for each unique idvar \code{subid}, and \code{"(1 + time|subid)"} fits a random intercept and random slope for each unique \code{subid}.}

\item{data}{name of the data frame that contains the variables above. Rows are different observations (subject/timepoints), columns are different variables.}

\item{adjustBatch}{should residuals be adjusted for the fixed effect of batch? Logical \code{TRUE} or \code{FALSE}. Use \code{FALSE} to illustrate additive (and multiplicative) batch effects. Use \code{TRUE} to illustrate only multiplicative batch effects. Default is \code{FALSE}.}

\item{orderby}{\code{'mean'} orders boxplots by increasing mean; best for illustrating additive batch effects (use with \code{adjustBatch=FALSE}). \code{'var'} orders boxplots by increasing variance; best for illustrating multiplicative batch effects. Default is \code{'mean'}.}

\item{plotMeans}{should batch means be plotted on top of the boxplots? Logical \code{TRUE} or \code{FALSE}. Default is \code{TRUE}.}

\item{colors}{vector of colors the same length and order as \code{levels(as.factor(data[,batchvar]))} that determines the colors of the boxplots (character string of color names or hexadecimal codes). Default is \code{"grey"} for all.}

\item{xlabel}{x-axis label (character string). Default is \code{'batch'}.}

\item{ylabel}{y-axis label (character string). Default is \code{'residuals'}.}

\item{title}{main title for the plot, default is no title (character string).}

\item{verbose}{prints messages. Logical \code{TRUE} or \code{FALSE}. Default is \code{TRUE}.}

\item{...}{other graphical parameter arguments passed to \code{\link[graphics]{par}}.}
}
\value{
Creates a boxplot.
}
\description{
\code{batchBoxplot} function will plot residuals of linear mixed effects model for a single feature by batch to visualize additive and multiplicative batch effects. Data should be in "long" format. Depends on \code{lme4} package.
}
