% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchTimeViz.R
\name{batchTimeViz}
\alias{batchTimeViz}
\title{Visualize Batches Over Time}
\usage{
batchTimeViz(
  batchvar,
  timevar,
  data,
  xlabel = "time",
  ylabel = "batch",
  title = "",
  verbose = TRUE,
  ...
)
}
\arguments{
\item{batchvar}{character string that specifies name of the batch variable. Batch variable should be a factor.}

\item{timevar}{character string that specifies name of numeric variable that distinguishes within-subject repeated measures, e.g., time, age, or visit. Will be plotted along x-axis.}

\item{data}{name of the data frame that contains the variables above. Rows are different observations (subject/timepoints), columns are different variables.}

\item{xlabel}{x-axis label (character string). Default is \code{'time'}.}

\item{ylabel}{y-axis label (character string). Default is \code{'batch'}.}

\item{title}{main title for the plot (character string). Default is no title.}

\item{verbose}{prints messages. Logical \code{TRUE} or \code{FALSE}. Default is \code{TRUE}.}

\item{...}{other graphical parameter arguments passed to \code{\link[graphics]{par}}.}
}
\value{
Creates a plot.
}
\description{
\code{batchTimeViz} is a simple function that will visualize batches over time for multi-batch longitudinal data. Data should be in "long" format.
}
