% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longCombat.R
\name{longCombat}
\alias{longCombat}
\title{Harmonize Multi-batch Longitudinal Data}
\usage{
longCombat(
  idvar,
  timevar,
  batchvar,
  features,
  formula,
  ranef,
  data,
  niter = 30,
  method = "REML",
  verbose = TRUE
)
}
\arguments{
\item{idvar}{character string that specifies name of ID variable. ID variable can be factor, numeric, or character.}

\item{timevar}{character string that specifies name of numeric variable that distinguishes within-subject repeated measures, e.g., time, age, or visit.}

\item{batchvar}{character string that specifies name of the batch variable. Batch variable should be a factor.}

\item{features}{character string that specifies names of the numeric feature variables, or the numeric indices of the corresponding columns.}

\item{formula}{character string representing all fixed effects on the right side of the formula for the linear mixed effects model. This should be in the notation used by \code{lme4} and include covariates, time, and any interactions. For example, \code{"age + sex + diagnosis*time"} fits model with fixed effects age, sex, diagnosis, time, and the diagnosis*time interaction. Formula should NOT include batchvar and should NOT include random effects.}

\item{ranef}{character string representing formula for the random effects in the notation used by \code{lme4}. For example, \code{"(1|subid)"} fits a random intercept for each unique idvar \code{subid}, and \code{"(1 + time|subid)"} fits a random intercept and random slope for each unique \code{subid}.}

\item{data}{name of the data frame that contains the variables above. Rows are different observations (subject/timepoints), columns are different variables.}

\item{niter}{number of iterations for empirical Bayes step. Usually converges quickly in less than 30 iterations. Default is 30.}

\item{method}{method for estimating sigma in standardization step (character string). \code{'REML'} (default, more conservative type I error control) or \code{'MSR'} (more powerful, less conservative type I error control).}

\item{verbose}{prints messages. Logical \code{TRUE} or \code{FALSE}. Default is \code{TRUE}.}
}
\value{
Function outputs a list including the following:
\describe{
\item{\code{data_combat}}{data frame with columns idvar, timevar, and ComBat-harmonized data for each feature}
\item{\code{gammahat}}{data frame containing mean of standardized data for each batch (row) and feature (column)}
\item{\code{delta2hat}}{data frame containing variance of standardized data for each batch (row) and feature (column)}
\item{\code{gammastarhat}}{data frame containing empirical Bayes estimate of additive batch effects}
\item{\code{delta2starhat}}{data frame containing empirical Bayes estimate of multiplicative batch effects}
}
}
\description{
\code{longCombat} function will implement longitudinal ComBat harmonization for multi-batch longitudinal data. Longitudinal ComBat uses an empirical Bayes method to harmonize means and variances of the residuals across batches in a linear mixed effects model framework. Detailed methods are described in the manuscript at \url{https://www.biorxiv.org/content/10.1101/868810v4}. This is a modification of the ComBat function code from the \code{sva} package that can be found at \url{https://bioconductor.org/packages/release/bioc/html/sva.html} and \code{combat.R} that can be found at \url{https://github.com/Jfortin1/ComBatHarmonization}. Data should be in "long" format. Depends on \code{lme4} package.
}
