% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multTest.R
\name{multTest}
\alias{multTest}
\title{Test for Multiplicative Batch Effects}
\usage{
multTest(idvar, batchvar, features, formula, ranef, data, verbose = TRUE)
}
\arguments{
\item{idvar}{name of ID variable (character string).}

\item{batchvar}{name of the batch/site/scanner variable (character string).}

\item{features}{vector of names of the feature variables (character string) or the numeric indices of the corresponding columns.}

\item{formula}{character string representing everything on the right side of the formula for the model, in the notation used by \code{lme4} including covariates, time, and any interactions, e.g., \code{"age + sex + diagnosis*time"} fits model with main effects age, sex, diagnosis, and time and the diagnosis*time interaction. Formula should NOT include batchvar and should NOT include random effects.}

\item{ranef}{character string representing formula for the random effects in the notation used by lme4, e.g., \code{"(1|subid)"} fits a random intercept for each unique idvar \code{"subid"}, and \code{"(1 + time|subid)"} fits a random intercept and slope unique \code{"subid"}.}

\item{data}{name of the data.frame that contains the variables above. Rows are different subject/timepoints (long format), columns are different variables.}

\item{verbose}{prints messages (logical \code{TRUE} or \code{FALSE}).}
}
\value{
A data frame of Fligner-Killeen test results for each feature.
}
\description{
\code{multTest} function will test (Fligner-Killeen method) for multiplicative batch effects in the residuals for each feature after fitting a linear mixed effects model. Data should be in "long" format. Depends on \code{lme4} package.
}
