% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchBoxplot.R
\name{batchBoxplot}
\alias{batchBoxplot}
\title{Boxplot for Batch Effects}
\usage{
batchBoxplot(
  idvar,
  batchvar,
  feature,
  formula,
  ranef,
  data,
  adjustBatch = FALSE,
  orderby = "mean",
  plotMeans = TRUE,
  colors = "grey",
  xlabel = "batch",
  ylabel = "residuals",
  title = "",
  verbose = TRUE,
  ...
)
}
\arguments{
\item{idvar}{name of ID variable (character string).}

\item{batchvar}{name of the batch/site/scanner variable (character string).}

\item{feature}{name of the feature variable (character string) or the numeric index of the corresponding column.}

\item{formula}{character string representing everything on the right side of the formula for the model, in the notation used by \code{lme4} including covariates, time, and any interaction, e.g., \code{"age + sex + diagnosis*time"} fits model with main effects age, sex, diagnosis, and time and the diagnosis*time interaction. Formula should NOT include batchvar and should NOT include random effects.}

\item{ranef}{character string representing formula for the random effects in the notation used by \code{lme4}, e.g., \code{"(1|subid)"} fits a random intercept for each unique idvar \code{subid}, and \code{"(1 + time|subid)"} fits a random intercept and slope for unique \code{subid}.}

\item{data}{name of the data frame that contains the variables above. Rows are different subject/timepoints (long format), columns are different variables.}

\item{adjustBatch}{should residuals be adjusted for batch? (logical \code{TRUE} or \code{FALSE}). Use \code{FALSE} to illustrate additive (and multiplicative) batch effects. Use \code{TRUE} to illustrate only multiplicative batch effects. Default is \code{FALSE}.}

\item{orderby}{\code{'mean'} orders boxplots by increasing mean; best for illustrating additive batch effects (use with \code{adjustBatch=FALSE}). \code{'var'} orders boxplots by increasing variance; best for illustrating multiplicative batch effects.}

\item{plotMeans}{Should batch means be plotted on top of the boxplots? (logical \code{TRUE} or \code{FALSE}). Default is \code{TRUE}.}

\item{colors}{Vector of colors the same length and order as \code{levels(as.factor(data[,batchvar]))} that determines the colors of the boxplots (character string of color names or hexadecimal codes). Default is \code{"grey"}.}

\item{xlabel}{x-axis label, default is \code{'batch'} (character string).}

\item{ylabel}{y-axis label, default is \code{'residuals'} (character string).}

\item{title}{main title for the plot, default is no title (character string).}

\item{verbose}{prints messages (logical \code{TRUE} or \code{FALSE}). Default is \code{TRUE}.}

\item{...}{other graphical parameter arguments passed to \code{par()}.}
}
\value{
Creates a boxplot.
}
\description{
\code{batchBoxplot} function will plot residuals of linear mixed effect model for a simgle feature by batch to visualize additive and multiplicative batch effects. Data should be in "long" format. Depends on \code{lme4} packages.
}
