% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchTimeViz.R
\name{batchTimeViz}
\alias{batchTimeViz}
\title{Visualize Batches Over Time}
\usage{
batchTimeViz(
  batchvar,
  timevar,
  xlabel = "time",
  ylabel = "batch",
  title = "",
  data,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{batchvar}{name of the batch/site/scanner variable (character string).}

\item{timevar}{name of the time variable, e.g. age or time from baseline (character string).}

\item{xlabel}{x-axis label, default is \code{'time'} (character string).}

\item{ylabel}{y-axis label, default is \code{'batch'} (character string).}

\item{title}{main title for the plot, default is no title (character string).}

\item{data}{name of the data frame that contains the variables above rows are different subject/timepoints (long format), columns are different variables.}

\item{verbose}{prints messages (logical \code{TRUE} or \code{FALSE}). Default is \code{TRUE}.}

\item{...}{other graphical parameter arguments passed to \code{par()}.}
}
\value{
Creates a plot.
}
\description{
\code{batchTimeViz} is a simple function that will visualize batches over time for multi-batch longitudinal data. Data should be in "long" format.
}
