% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multTest.R
\name{multTest}
\alias{multTest}
\title{Test for Multiplicative Batch Effects}
\usage{
multTest(idvar, batchvar, features, formula, ranef, data, verbose = TRUE)
}
\arguments{
\item{idvar}{character string that specifies name of ID variable. ID variable can be factor, numeric, or character.}

\item{batchvar}{character string that specifies name of the batch variable. Batch variable should be a factor.}

\item{features}{character string that specifies names of the numeric feature variables, or the numeric indices of the corresponding columns.}

\item{formula}{character string representing all fixed effects on the right side of the formula for the linear mixed effects model. This should be in the notation used by \code{lme4} and include covariates, time, and any interactions. For example, \code{"age + sex + diagnosis*time"} fits model with fixed effects age, sex, diagnosis, time, and the diagnosis*time interaction. Formula should NOT include batchvar and should NOT include random effects.}

\item{ranef}{character string representing formula for the random effects in the notation used by \code{lme4}. For example, \code{"(1|subid)"} fits a random intercept for each unique idvar \code{subid}, and \code{"(1 + time|subid)"} fits a random intercept and random slope for each unique \code{subid}.}

\item{data}{name of the data frame that contains the variables above. Rows are different observations (subject/timepoints), columns are different variables.}

\item{verbose}{prints messages. Logical \code{TRUE} or \code{FALSE}. Default is \code{TRUE}.}
}
\value{
A data frame of Fligner-Killeen test results for each feature.
}
\description{
\code{multTest} function will test for multiplicative batch effects in the residuals for each feature after fitting a linear mixed effects model. Uses Fligner-Killeen method for significance testing. Data should be in "long" format. Depends on \code{lme4} package.
}
