% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajPlot.R
\name{trajPlot}
\alias{trajPlot}
\title{Plot Individual Subject Trajectories}
\usage{
trajPlot(
  idvar,
  timevar,
  feature,
  batchvar,
  point.shape = NULL,
  point.col = NULL,
  line.col = NULL,
  xlabel = "time",
  ylabel = "feature",
  title = "",
  xlimits = NULL,
  ylimits = NULL,
  margins = c(5, 5, 3, 1),
  data,
  subset = rep(TRUE, nrow(data)),
  verbose = TRUE,
  ...
)
}
\arguments{
\item{idvar}{name of ID variable (character string).}

\item{timevar}{name of variable that distinguishes within-subject repeated measures, e.g., time, age, or visit (character string). Will be plotted along x-axis.}

\item{feature}{name of the feature variable (character string) to be plotted over time, or the numeric index of the corresponding column. Will be plotted along y-axis.}

\item{batchvar}{name of the batch/site/scanner variable (character string). Will determine the shape of plotting points.}

\item{point.shape}{vector that encodes the point shape for the batch variable levels. Should be in the same order as \code{levels(data[subset, batchvar])}. See the \code{pch} argument in \code{?points}. Default cycles through \code{pch=c(1:25, 0)} by batch level.}

\item{point.col}{vector that encodes colors of the point for each observation (character string of color names or hexadecimal codes). Length should be equal to number of unique observations included (\code{nrow(data[subset,])}) and in the same order as in the data frame. Default is \code{'black'} for all.}

\item{line.col}{vector that encodes colors of the line for each subject (character string of color names or hexadecimal codes). Length should be equal to number of unique subjects included (\code{length(unique(data[subset,idvar]))}) and in the same order as in the data frame. Default is \code{'black'} for all.}

\item{xlabel}{x-axis label, default is \code{'time'} (character string).}

\item{ylabel}{y-axis label, default is \code{'feature'} (character string).}

\item{title}{main title for the plot, default is no title (character string).}

\item{xlimits}{two dimensional numeric vector giving minimum and maximum x-axis values.}

\item{ylimits}{two dimensional numeric vector giving minimum and maximum y-axis values.}

\item{data}{name of the data frame that contains the variables above. Rows are different observations (subject/timepoints), columns are different variables.}

\item{subset}{logical or numeric 0/1 vector the same length as \code{nrow(data)} denoting which observations (i.e., rows) to plot. Default is to include all rows.}

\item{verbose}{prints messages (logical \code{TRUE} or \code{FALSE}). Default is \code{TRUE}.}

\item{...}{other graphical parameter arguments passed to \code{par()}.}
}
\value{
Creates a plot.
}
\description{
\code{trajPlot} will plot individual subject trajectories for multi-batch longitudinal data. Each line represents a trajectory of observations of a single feature (e.g., left fusiform cortical thickness) over time for an individual subject. Plotting points are coded by batch. Data should be in "long" format.
}
